% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIF_functions.R
\name{KIFall}
\alias{KIFall}
\title{Kendall Interaction Filter (KIF)}
\usage{
KIFall(Sample, y, ncor = 2, threshold)
}
\arguments{
\item{Sample}{a numeric matrix containing the data with \eqn{n} rows (observations) and \eqn{p} columns (features).}

\item{y}{the response variable. \code{y} must be a factor.}

\item{ncor}{number of cores to use for code parallelization. Default is 2.}

\item{threshold}{the number of relevant couples to select.}
}
\value{
\code{KIFall} function returns the first "threshold" selected couples among all the selected couples.
}
\description{
an interaction screening procedure based on Kendall's Tau.
}
\details{
KIF measure aims to select the relevant couples for the multi-class classification task. It is based on Kendall's Tau and ranges form 0 to 1. KIF is a model-free procedure, handles continuous, categorical or a mixture of continuous-categorical features, and is invariant under monotonic transformations. \code{KIFall} function computes KIF scores of all possible couples, then returns the "threshold" couples with highest scores.
}
\examples{
\dontrun{
############### KIFall function ###############
library(mvtnorm)
set.seed(1)
n1 <- 100
n2 <- 100
n <- n1 +n2
p <- 500
sigma <- diag(p)
sigma[upper.tri(sigma)] <- 0.2
sigma[lower.tri(sigma)] <- 0.2
sigma1 <- sigma
sigma2 <- sigma
sigma1[1,2] <- 0.8
sigma1[2,1] <- 0.8
sigma1[3,4] <- 0.8
sigma1[4,3] <- 0.8
sigma2[3,4] <- -0.8
sigma2[4,3] <- -0.8
mean1 <- c(rep(0,p))
mean2 <- c(rep(0,p))
Sample <- rbind(rmvnorm(n1, mean1, sigma1), rmvnorm(n2, mean2, sigma2))
y <- c(rep(1,n1), rep(0,n2))
out <- KIFall(Sample, y, 4, 10)
}
}
